% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{clusterApply_hop}
\alias{clusterApply_hop}
\title{Run clusterApply on local sub-clusters created on the selected hosts.}
\usage{
clusterApply_hop(n, cl, x, fun, ...)
}
\arguments{
\item{n}{integer; size of the local socket sub-cluster to create.}

\item{cl}{cluster object; cluster on which the local sub-clusters will be executed.}

\item{x}{vector; first input argument of the `fun' function}

\item{fun}{function; function to be executed on the sub-clusters}

\item{...}{extra input arguments of the `fun' function}
}
\value{
List containing function outputs
}
\description{
Run clusterApply on local sub-clusters created on the selected hosts.
}
\examples{

# Extract information on allocated nodes for the user "username". This is hereinafter called the main cluster.
# cl.info <- baseutils::lscluster(user="username")

#Create and register the main cluster
# cl <- baseutils::regcluster(cl.info$nodes,cl.info$nodelist)

# Run the function `fun_cluster' on sub-clusters of 20 nodes that will be created on each nodes of the main cluster.
# Extra arguments `argument1' and `argument2' are two arguments of the `fun_cluster' function.
# argument1 is a vector whose elements will be subset into 20 equal chunks that will be passed to each node of the main cluster.
# null <- baseutils::clusterApply_hop(n = 20, cl = cl, x = argument1, fun = fun_cluster, argument2)

# De-register the main cluster
# snow::stopCluster(cl)

}
