% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.r
\name{create.merged.sql}
\alias{create.merged.sql}
\title{Attach all sqlite3 files from a given directory into a single file with the name of that directory}
\usage{
create.merged.sql(ld.in, tablename, overwrite = FALSE)
}
\arguments{
\item{ld.in}{: list of directories containing merged files}

\item{tablename}{: name of the table to be merged. All others will be dropped.}

\item{overwrite}{: overwrite merged file, if exists}
}
\description{
Attach all sqlite3 files from a given directory into a single file with the name of that directory
}
