% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqsplit.R
\name{eqsplit}
\alias{eqsplit}
\title{Split a vector into a list of equal number of elements}
\usage{
eqsplit(vec, n)
}
\arguments{
\item{vec}{vector to be split}

\item{n}{integer; size of the split subset}
}
\value{
A list of vectors of size `n' (except for last one, possibly) containing
all elements of `vec'.
}
\description{
Split a vector into a list of equal number of elements
}
