% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{fstopCluster}
\alias{fstopCluster}
\title{Force stop a cluster
Use the snow::stopCluster function and force kill all active jobs by user on the node}
\usage{
fstopCluster(cl)
}
\description{
Force stop a cluster
Use the snow::stopCluster function and force kill all active jobs by user on the node
}
