% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{ldply_hop}
\alias{ldply_hop}
\title{Run ldply on local sub-clusters created on the selected hosts.}
\usage{
ldply_hop(n, cl, x, fun, ...)
}
\arguments{
\item{n}{integer; size of the local socket sub-cluster to create.}

\item{cl}{cluster object; cluster on which the local sub-clusters will be executed.}

\item{x}{vector; first input argument of the `fun' function}

\item{fun}{function; function to be executed on the sub-clusters}

\item{...}{extra input arguments of the `fun' function}
}
\value{
List containing function outputs
}
\description{
Run ldply on local sub-clusters created on the selected hosts.
}
