% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{localcluster}
\alias{localcluster}
\title{Run a function on a local cluster.
Can be used to embed a local cluster into a remote one.}
\usage{
localcluster(x_local, fun_local, ...)
}
\arguments{
\item{x_local}{vector; input parameters to be used by snow::clusterApply.
This corresponds to the first argument of `fun_local'.
The length of x_local determines the size of the cluster that will be registered.}

\item{fun_local}{function that will be executed on the local cluster.}

\item{...}{extra arguments to be used by fun_local}
}
\value{
List containing the function outputs
}
\description{
Run a function on a local cluster.
Can be used to embed a local cluster into a remote one.
}
