% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{logfile}
\alias{logfile}
\title{Standardized log file output}
\usage{
logfile(logtext, fn, lev = 0, dots = FALSE, newline = FALSE, append = TRUE)
}
\arguments{
\item{logtext}{string, text to be exported in the log file}

\item{fn}{string, name of the log file}

\item{lev}{integer, indenting space to be added before `logtext'.
Default is 0.}

\item{dots}{logical, indicates if an ellipsis should be added at the end of `logtext'.
Default is FALSE.}

\item{newline}{logical, indicates if a new line should be added after `logtext'.
Default is FALSE.}

\item{append}{logical, indicates if the log file should be appended, otherwise an existing file would be replaced.
Default is TRUE.}
}
\description{
Standardized log file output
}
