% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{lscluster}
\alias{lscluster}
\title{List information on requested clusters.
Currently only works for clusters using a SLURM management system.}
\usage{
lscluster(user, allocated = TRUE)
}
\arguments{
\item{user}{character; name of the cluster user.}

\item{allocated}{logical; indicates of the cluster information should be limited to allocated (running) nodes.
Default is TRUE.}
}
\value{
a data frame containing the job ID, partition, username, status, node host names and node numbers associated with user.
}
\description{
List information on requested clusters.
Currently only works for clusters using a SLURM management system.
}
