% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{package}
\alias{package}
\title{Create or install package}
\usage{
package(name, check = FALSE)
}
\arguments{
\item{name}{package name}

\item{check}{run devtools::check on the package (default is FALSE)}
}
\description{
Create or install package
}
