% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.R
\name{rds}
\alias{rds}
\title{Read and write rds files}
\usage{
rds(name, data = NULL, dir.save = NULL, overwrite = FALSE)
}
\arguments{
\item{name}{name of the output rds file}

\item{data}{data to be written. If NULL, the rds file is loaded.}

\item{dir.save}{repository where the rds file is written}

\item{overwrite}{overwrite if existing file (default is FALSE)}
}
\description{
Read and write rds files
}
