% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{regcluster}
\alias{regcluster}
\title{Create and register a cluster; Cluster ID is returned.}
\usage{
regcluster(ncores = 1, names = "localhost", outfile = "/dev/null")
}
\arguments{
\item{ncores}{number of cpu per cluster}

\item{names}{clustaer names}

\item{outfile}{outfile for cluster runs}
}
\description{
Create and register a cluster; Cluster ID is returned.
}
