% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.r
\name{SQLite.attach}
\alias{SQLite.attach}
\title{Attach multiple sqlite3 files into a single sqlite3 file}
\usage{
SQLite.attach(lf.in, fn.out, tablename)
}
\arguments{
\item{lf.in}{: list of input sqlite3 files to be merged}

\item{fn.out}{: name of the output sqlite3 file}

\item{tablename}{: name of the table to be merged. All others will be dropped.}
}
\description{
Attach multiple sqlite3 files into a single sqlite3 file
}
